:recycle: 스테이지 모드의 테스트 코드에서 세부 기록까지 StageDefenseRecordTest에서 검증하도록 수정
