+++ src/main/java/kr/co/morandi/backend/domain/contentproblemrecord/random/StageDefenseProblemRecord.java
@@ -32,7 +32,7 @@ public class StageDefenseProblemRecord extends ContentProblemRecord {
     public static StageDefenseProblemRecord create(Member member, Problem problem,
                                                    ContentRecord contentRecord, ContentType contentType) {
         List<ContentProblemRecord> contentProblemRecords = contentRecord.getContentProblemRecords();
-        Long stageNumber = (long) ((contentProblemRecords == null) ? 0 : contentProblemRecords.size() + 1);
+        Long stageNumber = (long) ((contentProblemRecords == null) ? 1L : contentProblemRecords.size() + 1);
         return new StageDefenseProblemRecord(stageNumber, member, problem, contentRecord, contentType);
     }
 }
